% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_bandizer.R
\name{age_bandizer}
\alias{age_bandizer}
\title{Age Band Creation: Create a new column of 5 year Age Bands from an integer column}
\usage{
age_bandizer(df, Age_col)
}
\arguments{
\item{df}{a tidy dataframe in standard Master Patient Index format ie SangerTools::PopHealthData}

\item{Age_col}{a integer column within @param df NAs must be removed or imputed prior to running this function}
}
\value{
A dataframe with width ncol(df)+1, new column will be named Ageband and will be a factor with levels defined
}
\description{
Age Band Creation: Create a new column of 5 year Age Bands from an integer column
}
\examples{
library(SangerTools)
library(dplyr)
health_data <- SangerTools::PopHealthData
}
