% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_brand_palette.R
\name{show_brand_palette}
\alias{show_brand_palette}
\title{Core brand colour palette}
\usage{
show_brand_palette()
}
\value{
A character vector of length 5 containing the hex codes for the
core palette.
}
\description{
Displays the five core SangerTools brand colours drawn from NHS Identity
guidelines, alongside their hex codes.
}
\examples{
library(SangerTools)
show_brand_palette()
}
