% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_save.R
\name{split_and_save}
\alias{split_and_save}
\title{Split & Save}
\usage{
split_and_save(df, Split_by, path, prefix = NULL)
}
\arguments{
\item{df}{A `dataFrame` or `tibble` ie PopHealthData.}

\item{Split_by}{A column within df that will be used to split the patients and will also appear in the file name. Ideally should be a health organisation code such as GP Practice Code or Hospital Trust Code. Should only have alpha-numeric values}

\item{path}{A file path to which the CSV files will be written}

\item{prefix}{File name prefix}
}
\value{
n number of CSV files written to the location specified by path argument.
}
\description{
A simpler alternative to  \code{\link{cohort_processing}}. Will split a data frame
and save as a csv
}
\examples{
 \dontrun{
split_and_save(
 df = pseudo_data,
 Split_by = "Locality",
 file_path = "Inputs/",
 prefix = NULL
)
}



}
