% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardised_rates.R
\name{standardised_rates}
\alias{standardised_rates}
\title{Direct Standardised Prevalence Rates}
\usage{
standardised_rates(
  df,
  Geography,
  LocalityA,
  LocalityB,
  Condition,
  Population_Standard = NULL,
  Verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{a tidy dataframe in standard Master Patient Index format ie SangerTools::PopHealthData}

\item{Geography}{Column name denoting column with location string names}

\item{LocalityA}{A character string within @param Geography must be passed within ""}

\item{LocalityB}{A character string with @param Geography must be passed within ""}

\item{Condition}{A Health condition flag denoted by 1 & 0}

\item{Population_Standard}{Population Standard Weight for Standardising; defaults to NULL which will use Age Structure of @param  df}

\item{Verbose}{Takes boolean value to determine function output printing}

\item{...}{Variables to Standardise by; Must always have Ageband, additional variables are optional}
}
\value{
List of length 2 containing 2 tibbles with standardised rates for @param LocalityA and @param LocalityB
}
\description{
Direct Standardised Prevalence Rates
}
