% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_pop_standard.R
\docType{data}
\name{uk_pop_standard}
\alias{uk_pop_standard}
\title{Data set of 2018 UK Population}
\format{
A tibble with 29 rows and 2 variables:
\describe{
  \item{UK_Population}{dbl Year price was recorded}
  \item{Ageband}{5 Year age band for population}
  }
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates}
}
\usage{
uk_pop_standard
}
\description{
Data is taken from ONS and is split into 5 year age band
}
\keyword{datasets}
