% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_col_chart.R
\name{categorical_col_chart}
\alias{categorical_col_chart}
\title{Plot Counts of Categorical Variables
Create a ggplot2 column chart of categorical variables with proportions, in ascending order.}
\usage{
categorical_col_chart(df, grouping_var)
}
\arguments{
\item{df}{A dataframe with categorical variables}

\item{grouping_var}{a categorical variable by which to group the count by}
}
\value{
a ggplot2 object
}
\description{
Plot Counts of Categorical Variables
Create a ggplot2 column chart of categorical variables with proportions, in ascending order.
}
\examples{
library(SangerTools)
library(gapminder)
# Generate a plot using the gap minder dataset
plot <- categorical_col_chart(gapminder,continent)
print(plot)
}
