% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude_rates.R
\name{crude_rates}
\alias{crude_rates}
\title{Crude Prevalence Calculator}
\usage{
crude_rates(df, Condition, ...)
}
\arguments{
\item{df}{a tidy dataframe in standard Master Patient Index format ie SangerTools::PopHealthData}

\item{Condition}{A Health condition flag denoted by 1 & 0}

\item{...}{Grouping variables ie Locality; use at least 1, when multiple seperate by comma}
}
\value{
a tibble Rates Per 1,000 calculated
}
\description{
Crude Prevalence Calculator
}
\examples{
library(SangerTools)
library(dplyr)
health_data <- SangerTools::PopHealthData
glimpse(health_data)
# Generate crude prevalene rate stats
crude_prevalence <- SangerTools::crude_rates(health_data,Diabetes,Locality)
print(crude_prevalence)
}
