% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_sql.R
\name{df_to_sql}
\alias{df_to_sql}
\title{DataFrame to SQL; Write your DataFrame or Tibble directly to SQL from R}
\usage{
df_to_sql(df, driver, server, database, sql_table_name, ...)
}
\arguments{
\item{df}{A DataFrame or Tibble ie PopHealthData}

\item{driver}{A driver for database ie "SQL Server"; must be passed in quotation}

\item{server}{The unique name of your database server; must be passed in quotation}

\item{database}{The name of the database to which you will write `df`; must be passed in quotation}

\item{sql_table_name}{The name that `df` will be referred to in SQL database; must be passed in quotation}

\item{...}{Function forwarding for additional functionality}
}
\value{
A message confirming that a new  table has been created in a SQL `database`
}
\description{
DataFrame to SQL; Write your DataFrame or Tibble directly to SQL from R
}
\examples{
\dontrun{
library(odbc)
library(DBI)
health_data <- SangerTools::PopHealthData
df_to_sql(dataframe = health_data,
            driver  = "SQL SERVER",
           database = "DATABASE",
     sql_table_name = "New Table Name")
}
}
