% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_clip.R
\name{excel_clip}
\alias{excel_clip}
\title{Dataframe to Clipboard
This function copies a dataframe to your clipboard.}
\usage{
excel_clip(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Dataframe}

\item{row.names}{Set to FALSE}

\item{col.names}{Set to TRUE}

\item{...}{function forwarding}
}
\value{
a data frame copied to your clipboard
}
\description{
Dataframe to Clipboard
This function copies a dataframe to your clipboard.
}
\examples{
\dontrun{
excel_clip(dataframe)}
}
