% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_clip.R
\name{excel_clip}
\alias{excel_clip}
\title{Dataframe to Clipboard
This function copies a data frame to your clipboard.}
\usage{
excel_clip(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Dataframe}

\item{row.names}{Set to FALSE for row.names not to be included}

\item{col.names}{Set to TRUE for col.names to be included}

\item{...}{function forwarding for additional write.table functionality}
}
\value{
a data frame copied to your clipboard
}
\description{
Dataframe to Clipboard
This function copies a data frame to your clipboard.
}
\examples{
\donttest{
library(SangerTools)
library(dplyr)
# Group by Age Band
health_data <- SangerTools::PopHealthData
health_data \%>\%
 slice(1:10) \%>\%
 SangerTools::excel_clip(row.names = FALSE, col.names = TRUE)
 # Message: Your Item is Now Copied to Clipboard
 }
}
