% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_excel_reader.R
\name{multiple_excel_reader}
\alias{multiple_excel_reader}
\title{Read Multiple Excel files into R
This function reads multiple Xlsx files in a directory must be same structure.}
\usage{
multiple_excel_reader(file_path, rows_to_skip = 0, col_names = TRUE)
}
\arguments{
\item{file_path}{The Directory in which the files are located}

\item{rows_to_skip}{The number of rows from the top to be excluded}

\item{col_names}{If columns are named; defaults to TRUE}
}
\value{
a data frame object full of file paths
}
\description{
Read Multiple Excel files into R
This function reads multiple Xlsx files in a directory must be same structure.
}
\examples{
\dontrun{
combined_excel_files <- multiple_excel_reader("Inputs/", 1, TRUE)
}
}
