% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_brand_palette.R
\name{show_brand_palette}
\alias{show_brand_palette}
\title{Brand Colour Palette}
\usage{
show_brand_palette()
}
\value{
a Base R plot object
}
\description{
Displays a brand colour palette for showing the hex codes associated with brand
}
\examples{
library(scales)
library(SangerTools)
show_brand_palette()
}
