% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_extended_palette.R
\name{show_extended_palette}
\alias{show_extended_palette}
\title{Extended Brand Colour Palette}
\usage{
show_extended_palette()
}
\value{
a Base R plot object
}
\description{
Displays extended brand colour palette for charting
}
\examples{
library(scales)
library(SangerTools)
show_extended_palette()
}
