% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardised_rates.R
\name{standardised_rates}
\alias{standardised_rates}
\title{Direct Standardised Prevalence Rates}
\usage{
standardised_rates(
  df,
  Geography,
  LocalityA,
  LocalityB,
  Condition,
  Population_Standard = NULL,
  Verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{a tidy dataframe in standard Master Patient Index format ie SangerTools::PopHealthData.}

\item{Geography}{Column name using NSE;passed without quotation denoting column where @param LocalityA and @param LocalityB are contained.}

\item{LocalityA}{A value within @param Geography using SE; must be passed within quotation.}

\item{LocalityB}{A value within @param Geography using SE; must be passed within quotation.}

\item{Condition}{A Health condition flag denoted by 1 & 0; where 1 denotes the patient being positive for the health condition.}

\item{Population_Standard}{Population Standard Weight used for Standardising; default set to NULL; which denotes  use of Age Structure of @param  df .}

\item{Verbose}{Takes a boolean value to determine function output printing. TRUE will print two dataframes however they will be different to the the output of the function.}

\item{...}{Variables used to standardise by; Must always have Ageband, additional variables are optional and should be passed separated by commas.}
}
\value{
List of length 2 containing 2 tibbles with standardised rates for @param LocalityA and @param LocalityB .
}
\description{
Ideal for providing standardised prevalence comparisons of two populations specified by `LocalityA` and `LocalityB`.
Message printing displays standardised  rates and a comparative between the two populations. The ratio
is presented as `LocalityA`:`LocalityB`,whereby a number greater than 1 signifies
`LocalityA` having a higher rate, while a number less than one signifies `LocalityA` having a lower
rate. Outputs printed via `Verbose = TRUE`, are different to outputs returned.
The returned output, a list of length 2,has column names cleaned by `janitor::clean_names()`
and a new name column corresponding to `LocalityA`and `LocalityB`. This was done
with downstream ggplot2 plotting in mind.
}
